Move files to shared module.
